UI components modularized.
